processing/requirements.txt
applauncher==2.1.0
APScheduler==3.9.1
apscheduler-bundle==2.0.0
attrs==22.1.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
clickclick==20.10.2
colorama==0.4.6
commonmark==0.9.1
connexion==2.14.1
dependency-injector==4.40.0
Flask==2.0.3
greenlet==1.1.3.post0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
packaging==21.3
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.28.1
rich==9.13.0
six==1.16.0
SQLAlchemy==1.4.42
swagger-ui-bundle==0.0.9
typer==0.3.2
typing-extensions==3.10.0.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
